shifted FindCommandNew to FindCommand
