+++ src/main/java/seedu/address/logic/commands/FindCommand.java
@@ -2,27 +2,33 @@ package seedu.address.logic.commands;
 
 import static java.util.Objects.requireNonNull;
 
+import java.util.Optional;
+import java.util.function.Predicate;
+
 import seedu.address.commons.core.Messages;
 import seedu.address.logic.CommandHistory;
 import seedu.address.model.Model;
-import seedu.address.model.person.NameContainsKeywordsPredicate;
+import seedu.address.model.person.Grade;
+import seedu.address.model.person.Person;
+import seedu.address.model.person.Semester;
 
 /**
- * Finds and lists all persons in address book whose name contains any of the argument keywords.
- * Keyword matching is case insensitive.
+ * Finds and lists module(s) in module plan matching all given module code,
+ * semester taken or grade obtained (case-insensitive).
  */
 public class FindCommand extends Command {
 
     public static final String COMMAND_WORD = "find";
 
-    public static final String MESSAGE_USAGE = COMMAND_WORD + ": Finds all persons whose names contain any of "
-            + "the specified keywords (case-insensitive) and displays them as a list with index numbers.\n"
-            + "Parameters: KEYWORD [MORE_KEYWORDS]...\n"
-            + "Example: " + COMMAND_WORD + " alice bob charlie";
+    public static final String MESSAGE_USAGE = COMMAND_WORD + ": Finds and lists module(s) based on "
+            + "module code, semester taken or grade obtained (case-insensitive).\n"
+            + "Module code can be entered partially, but semester and grade must be exact.\n"
+            + "Parameters: [c/MODULE_CODE] [s/SEMESTER] [g/GRADE_OBTAINED]\n"
+            + "Example: " + COMMAND_WORD + " s/y1s1 c/cs g/A";
 
-    private final NameContainsKeywordsPredicate predicate;
+    private final Predicate<Person> predicate;
 
-    public FindCommand(NameContainsKeywordsPredicate predicate) {
+    public FindCommand(Predicate<Person> predicate) {
         this.predicate = predicate;
     }
 
@@ -39,5 +45,62 @@ public class FindCommand extends Command {
         return other == this // short circuit if same object
                 || (other instanceof FindCommand // instanceof handles nulls
                 && predicate.equals(((FindCommand) other).predicate)); // state check
+    }
+
+    /**
+     * Stores the details for finding a module. Each field can be empty.
+     */
+    public static class FindModuleDescriptor {
+        private String code; // can be substring of exact code
+        private Semester semester;
+        private Grade grade;
+
+        public FindModuleDescriptor() {}
+
+        /**
+         * Copy constructor.
+         */
+        public FindModuleDescriptor(FindModuleDescriptor toCopy) {
+            setCode(toCopy.code);
+            setSemester(toCopy.semester);
+            setGrade(toCopy.grade);
+        }
+
+        public void setCode(String code) {
+            this.code = code;
+        }
+
+        public Optional<String> getCode() {
+            return Optional.ofNullable(code);
+        }
+
+        public void setSemester(Semester semester) {
+            this.semester = semester;
+        }
+
+        public Optional<Semester> getSemester() {
+            return Optional.ofNullable(semester);
+        }
+
+        public void setGrade(Grade grade) {
+            this.grade = grade;
+        }
+
+        public Optional<Grade> getGrade() {
+            return Optional.ofNullable(grade);
+        }
+
+        @Override
+        public boolean equals(Object object) {
+            if (object == this) {
+                return true;
+            }
+            if (!(object instanceof FindModuleDescriptor)) {
+                return false;
+            }
+
+            FindModuleDescriptor other = (FindModuleDescriptor) object;
+            return (code.equals(other.code) && semester.equals(other.semester) && grade.equals(other.grade));
+        }
     }
 }
